refactor: move some components of reposition
