+++ packages/web/src/views/pool/pool-reposition/components/balance-change/BalanceChange.styles.ts
@@ -1,8 +1,10 @@
 import styled from "@emotion/styled";
+
 import { fonts } from "@constants/font.constant";
-import mixins from "@styles/mixins";
 import { media } from "@styles/media";
-import { RepositionContentBoxStyle } from "../reposition-content/RepositionContent.styles";
+import mixins from "@styles/mixins";
+
+import { RepositionContentBoxStyle } from "../reposition-content-box/RepositionContnetBox.style";
 
 interface Props {
   isDisabled?: boolean;
