+++ packages/web/src/views/pool/pool-reposition/components/reposition-content/RepositionContent.tsx
@@ -10,16 +10,16 @@ import {
 } from "@constants/option.constant";
 import { IPriceRange } from "@hooks/increase/use-increase-handle";
 import { SelectPool } from "@hooks/pool/use-select-pool";
-import { REPOSITION_BUTTON_TYPE } from "@hooks/reposition/use-reposition-handle";
 import { TokenAmountInputModel } from "@hooks/token/use-token-amount-input";
 import { PoolPositionModel } from "@models/position/pool-position-model";
 import { TokenModel } from "@models/token/token-model";
 
+import { REPOSITION_BUTTON_TYPE } from "../../hooks/use-reposition-handle";
 import BalanceChange from "../balance-change/BalanceChange";
 import RepositionSelectPosition from "../reposition-select-position/RepositionSelectPosition";
 import RepositionSelectRange from "../reposition-select-range/RepositionSelectRange";
-import { ToolTipContentWrapper } from "../reposition-select-range/RepositionSelectRange.styles";
-import { RepositionContentWrapper } from "./RepositionContent.styles";
+
+import { RepositionContentWrapper, ToolTipContentWrapper } from "./RepositionContent.styles";
 
 interface RepositionContentProps {
   tokenA: TokenModel | null;
